Move delta to be a purely server-side view
It has no dynamic/client-side elements, and is only included by other server-side views.
